debt: tests: split control-logic.test.js into per-concern files
2084 → 1130 (main) + 923 (new control-logic-valves.test.js). Both
under the 1200-line test cap.

Split at the planValveTransition boundary: first file keeps mode
evaluation, hysteresis, valve/actuator mapping, priorities, sensor
failure, emergency overlay, formatters, config-gated behaviour and
manual override. Second file takes planValveTransition,
runBoundedPool, buildSnapshotFromState, evaluate-reason tests, and
v_air polarity. assertInvariants + VALVE_NAMES + allClosed /
allOpenSinceZero helpers moved with them.

npm test:unit 782/782 pass.
